Fix #26: Implement Entity as HomeAssistant requires
