+++ custom_components/fallback_conversation/manifest.json
@@ -8,10 +8,11 @@
     "dependencies": [
       "conversation"
     ],
+    "after_dependencies": ["assist_pipeline"],
     "documentation": "https://github.com/m50/ha-fallback-conversation",
     "integration_type": "service",
     "iot_class": "local_polling",
     "issue_tracker": "https://github.com/m50/ha-fallback-conversation/issues",
     "requirements": [],
-    "version": "1.0.5"
+    "version": "1.1.0"
   }
